requirements.txt
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.34.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.0
gitdb==4.0.12
GitPython==3.1.41
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msrest==0.7.1
oauthlib==3.2.2
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.6.0
six==1.17.0
smmap==5.0.2
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
